+++ include/bitsery/details/not_defined_type.h
@@ -56,7 +56,7 @@ namespace bitsery {
             int& operator*() {
                 return data;
             }
-            int data;
+            int data{};
         };
 
         template <typename T>